datadog/samples/travel-concierge-agent-observability/concierge_agent/supervisor_agent/requirements.txt
bedrock-agentcore>=0.1.2
bedrock-agentcore-starter-toolkit>=0.1.6
boto3>=1.40.8
pydantic>=2.0.0
pydantic-core>=2.33.2
pytz>=2025.2
requests>=2.28.0
requests-aws4auth>=1.3.1
strands-agents[otel]>=1.4.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-otlp-proto-http>=1.20.0
mcp
